requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.10
faiss-cpu==1.8.0
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.2.2

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Data & Evaluation
datasets==2.15.0
pandas==2.1.3
numpy==1.26.4
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1